prepare courseview for lectures and attendances
