+++ packages/frontend/src/components/Layout.tsx
@@ -1,6 +1,6 @@
 export const Layout: React.FC<{ children: JSX.Element }> = ({ children }) => {
     return (
-        <div className="flex items-center justify-center min-h-screen">
+        <div className="flex items-top justify-center min-h-screen">
             {children}
         </div>
     );
